Fix unfixed pid_b batch-axis overflow in _chunk_state_varlen_kernel
Found while double-checking why _chunk_state_varlen_kernel was marked
batch-axis n/a in the coverage audit -- it actually does have a pid_b
term (states_ptr += pid_b * stride_states_batch + ...), just uncast, the
same missed pattern as ssd_bmm.py's pid_b and ssd_state_passing.py's
pid_m found earlier.

Unlike every other batch-axis case in this file, `states` here is
freshly allocated (ordinary contiguous) inside chunk_state_varlen()
itself -- its batch stride is just nheads * headdim * dstate, its own
natural size, not an injected wide stride. So this needs a genuinely
large *batch* (tens of thousands of packed sequences), not a synthetic
parent tensor -- realistic at continuous-batching serving scale, not just
a theoretical corner case.

Confirmed via direct reproduction before this fix: batch=65,000 packed
single-token sequences with nheads*headdim*dstate=36,864 crashes with an
illegal memory access. Cast pid_b to int64; the same repro now matches an
independent per-sequence recompute at several batch indices, including
right at the point the old code would have wrapped.

Third real crash bug found this way (after ssd_bmm's pid_b and
ssd_state_passing's pid_m) from a "should be fine by analogy" assumption
that turned out wrong under direct verification.

+++ mamba_ssm/ops/triton/ssd_chunk_state.py
@@ -659,13 +659,13 @@ def _chunk_state_varlen_kernel(
     # Meta-parameters
     BLOCK_SIZE_M: tl.constexpr, BLOCK_SIZE_N: tl.constexpr, BLOCK_SIZE_K: tl.constexpr,
 ):
-    pid_b = tl.program_id(axis=1)
+    # int64 to avoid int32 overflow, see TODO: LinkToFutureIssueInMamba
+    pid_b = tl.program_id(axis=1).to(tl.int64)
     pid_h = tl.program_id(axis=2)
     num_pid_n = tl.cdiv(dstate, BLOCK_SIZE_N)
     pid_m = tl.program_id(axis=0) // num_pid_n
     pid_n = tl.program_id(axis=0) % num_pid_n
     end_idx = tl.load(cu_seqlens_ptr + pid_b + 1)
-    # int64 to avoid int32 overflow, see TODO: LinkToFutureIssueInMamba
     pid_c = ((end_idx - 1) // chunk_size).to(tl.int64)
     b_ptr += pid_c * chunk_size * stride_b_seqlen + (pid_h // nheads_ngroups_ratio) * stride_b_head
     x_ptr += pid_c * chunk_size * stride_x_seqlen + pid_h * stride_x_head
